Random updates and cleanup, prep work for the next step
Kind of floundered here a bit and realized that it's time to really grok
the parser side of things and start figuring out how Late's parser will
work. Time to start throwing together a Pratt parser for actually
handling the liquid code.

Also, naming. Naming is hard. Going to try to stick to the "Expressions
produce values, Statements don't" nomenclature.