Infix and Prefix operator parsing!
This is cool. The precedence stack plus the Pratt Parsing algorithm and
all of this is parsed and handled in just a couple of lines of code.

Need to play around more with error handling and finalizing what syntax
is considered an error and what isn't (e.g. is empty `{{ }}` an error or just a
no-op?).

Also hack-fixing a panic in `make docs` due to the incompleteness of the
parser.