Lex and Parse boolean literals
This is our first case of an actual keyword that we support. Late will
have the minimal number of these, the rest should be filters or tags,
but in terms of usability making `true` and `false` built in is
expected.

Also did some test suite cleanup, building out a few helper functions.