Evaluate basic filters!
This required a couple of changes to how we're doing things. After
putting in the pieces in the parser to get contextual nodes so we know
we are dealing with a filter expression, figuring out where to store
filters, how to access them, and how to call them became the big
question. To make things easy, I'm starting with basic filters with no
parameters.

Due to the dynamic nature of Liquid, our object system needs to be just
as dynamic. This conflicts pretty hard with Go's type system, requiring
some mapping and wrapping layers to let us work with a more dynamic type
system on top of Go. With that mostly in place, I'm starting with
filters taking `object.Objects` though I'd love to play with making
filters straight-up go methods with types and building wrappers around
them for the object <-> Go types management. This will play in even
stronger when I play with C-exports and the FFI for defining custom
filters.