Fix parsing filter parameters and chained filters
Realized I didn't have a test for filters + params + more filters and
found out that it didn't work at all!

First, had to implement proper inspect output of FilterLiteral ast nodes
so I could see what my problem was, which showed me that the parser was
taking the rest of the full expression and building a new sub-expression
with filter applications and everything. Specifically, the test show the
following output:

  replace: "this", with: ("that" | upcase)

instead of what I expected:

  (replace: "this", with: "that") | upcase

So after racking my brain on how to trick the parser to know that it's
in filter-parameter mode and should quit at the next pipe it sees, I
realized this feature is already built into the parser: precedence!
Because the Pratt parser rules are `precendence < peekPrecedence` and
not `<=`, we set the minimum precendence to the level on which we want
to stop.

In the end, I had to change one token (`LOWEST` to `PIPE`) and not only
did it work perfectly, but now I have the ability to run nested filter
expressions as well, something that would be a nightmare to implement
in Liquid itself!