Now supports basic, global variables
First step in getting this as a proper templating language is the
ability to assign variables for the render context. This led to
requiring a new Context object when running a Template. For now the
Context will be the catch-all for all information regarding the
currently running render. That may change in the future but it's handy
for now.