Move Filter and Tag collections to the top level package
I don't think there's any reason to have a per-render-level set of
filters or tags at this point. This takes quite a bit of logic out of
pieces that really didn't need it (context, mainly) and sets up a very
obvious way to add new tags or filters to the system.

Tags have a special case that we want to make sure we're creating a new
tag object every time one is encountered during parsing. Thus, we
provide an API for setting up a generator method linked to a name to
ensure this is the case.