Implement the comment block
The comment block simply throws itself away and anything inside of the
block. This required creating the ability to parse a block of liquid
code and associate it with a specific tag, which led me to finally do
some refactoring of the lexer to let me create bigger tokens. Turning
the `{% end %}` sigil into a token makes it trivial to find in the
parser instead of treating it like its own tag.