Implement the 'raw' tag
This tag is a special case that needs to re-print the raw input for all
code inbetween {% raw %} ... {% end %}. This is the first block-type tag
I've implemented so still playing with ways of exposing a list of
statements to a tag to execute.

I had to remove the test-level inspection in the String() method of ast
node types because String() is now expected to print out the actual code
of a given node. As such I need to figure out a new way of writing the
oeprator precedence test, which was the only code using Node.String()
anyway.