Rebuild the Lexer to track whitespace
As Late's a templating language, one thing we need to worry about that a
regular language doesn't is re-building the original template verbatim
from its parts. The lexer now keeps track of a Raw and Literal value of
a token, Raw being the value including all leading and internal
whitespace, and Literal being a token value with all extraneous
whitespace removed.