Support Comments and Raw / Verbatim blocks
There's a lot going on here, partly due to me trying to figure out how
to make {% comment %} and {% raw %} blocks to work. These required
disabling parsing and I was trying really hard not to dive into the
lexer as it shouldn't know about tag implementations. I eventually
decided that, this is a new language, lets make these syntax rules and
not tags.

Comments are now `{# #}` and raw/verbatim blocks are `{{{ }}}`. This
means we can put logic into the Lexer to properly tokenize everything
inside of these tokens as just raw text, which also opens up the parser
to a much simpler implementation for these situations.

I'm keeping around the block-tag work here as it will come in handy as I
work on actual block tags like `if` and `for`.