Support dot access for deeper object/hash entries
This effectively adds two more kinds of use cases: accessing hash data
as well as nested objects, though underneath it's one and the same. Late
will be like Javascript in this sense, objects are just dictionaries of
stuff, and it's up to the implementors and users to define how the
values are handled.