Support the `if` tag
This was a doozy! Supporting `if` is more than just supporting a block
tag, I also had to set up a concept and system of SubTags because `if`
also needs to support `elsif` and `else` clauses, and those need to be
dynamically managed in the parser and the evaluator without knowing what
exactly is coming and when.

So this got kind of messy and I'll look at ways of cleaning up the tag
parsing code later. However I'm really happy at how the actual
implementation of the `if` tag turned out. I'm not sure if you can get
much simpler or more obvious than that!