Implement the basic `for` loop complete with `continue` and `break`
While the loop itself was simple enough to construct, figuring out a
solution for `continue` and `break` proved far more difficult, because
these are not considered "subtags" and can appear at any depth inside of
the for loop's block. These tags need to halt all further execution of
the for loop.

I first tried panic/recover, but that quickly ran into problems with
making available all statements that had already been evaluated up to
the point of the panic. Having recover for every single evaluated
statement would be not only slow but probably annoying for future
concerns, so I looked for a different route.

Having investigated other patterns for handling this kind of thing, I've
settled on a system that allows a tag to be an Interrupt. The parser
will see this tag and insert an InterruptStatement into the tree, and
the evaluator will halt all further evaluation and unwind itself back up
the tree once it sees this entry. Once an interrupt is cleared,
evaluation will continue automatically.