Make sure tag Expression() nodes evaluate variables
Realized here that I was trying to use Identifiers in two different
contexts when dealing with parsing tag nodes. When we're in the
evaluator, an Identifier should always try to eval as a variable and
return whatever is stored under that name. To work with the Identifier()
parse rule, that now gets turned into a StringLiteral instead, just like
Literal(), ensuring the evaluator can just return the value and move on.