Update docs script to work with individual Liquid test blocks
Having noticed that adding raw tags `{{{ }}}` to the text description of
sections in the docs means that the resulting output of files will never
fully match the generated templates, I decided to only run each
individual Liquid segment (the `>` and `<` lines) as their own
templates. This lets me better find out which sections work and which
don't.

Also trying a different way of showing failure diffs as the previous
tool wouldn't show whitespace diffs very well. The current shows it
slightly better but it's still not what I really want.