First pass of the include tag!
This got really messy and I'll be working to clean things up but the
basics now work. After a lot of thought and various tests on how to work
around an import cycle caused by the `include` tag importing `template`
to trigger a new render, I've come up with the following steps to
disconnect tags from the render process as much as possible.

* Inject the template's render function into the context
* Expose this to Tags through the Environment interface.
* Set up Context to be configured with a FileReader to decouple the file
system from rendering.
* Also expose the Context's file reader API to tags via Environment

The context configuration pattern is pretty cool. I found this idea of
functional options from Dave Cheney
(https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis)
that looks awesome and gives a nice, chainable way to create and provide
new options.