Support the `forloop` hash inside of a for loop
This object provides information regarding the for loop iteration that
is useful for template designers. I don't support everything Liquid
supports yet but more can easily be added as needed.

This brought about an interesting design decision. One thing that I
always get bit with Liquid is not understanding what scope I'm in. As
such Late tries to run with effectively 2 scopes: file-scope and
global-scope. With this, `assign` will always make a variable that the
entire current template can access. However, things like `forloop` and
the for loop's local iteration variable need to be scoped to live only
in the for loop's block.

This has led to the creation of what I'm calling a "Shadow" scope. You
can set values on the shadow and they will be lost on the next PopScope,
but any attempt to Set values normally will pass through the Shadow to
be caught by the first "real" scope on the stack.