Make more direct use of the Context
Having noticed that the "Environment" interface ended up being a
majority pass through to the Context, I'm able to get rid of this
"Environment" nomenclature and just lean on the Context for these kind
of hooks. The evaluator now injects itself into the context to expose
certain aspects of itself to tags, and otherwise the rest is just
handled by the Context directly.